alter: convert tuples of upgraded space on-the-fly
If upgrade of the space is in the progress, let's convert tuples on the
fly - they are not inserted in the space in case they haven't been
updated yet. Note that in this case some tuples are processed with upgrade
function twice. That's why upgrade function's idempotency is a vital
requirement.

Part of #xxxx