alter: refactor alter space lock
Alter lock was implemented as class, which maintains set of space ids to
be locked: on class construction lock is acquired (added to set) and on
destruction - removed. We are going to extend lock till the end of
upgrade process. So from now we can't release lock on destruction.
Hence, let's get rid of the alter lock class and use simple C-like
wrappers: alter_space_lock() and alter_space_unlock().

Part of #xxxx