iterator: introduce create_upgrade_iterator vtab method
Upgrade iterator is designed to return unmodified data during upgrade
(in contrast to casual iterator). To achieve this we simply raise flag
iterator->is_raw. Except for that - under the hood it is casual ITER_ALL
iterator.

Upgrade iterator is defined only for memtx tree index for now.

Needed for #xxxx